リファクタリングとモデル保存機能の追加
